+++ bin/generate.js
@@ -40,50 +40,54 @@ var E = __importStar(require("fp-ts/lib/Either"));
 var A = __importStar(require("fp-ts/lib/Array"));
 var pipeable_1 = require("fp-ts/lib/pipeable");
 var function_1 = require("fp-ts/lib/function");
-var utils_1 = require("./utils");
+var utils_1 = require("./services/utils");
 var type_guards_1 = require("./type-guards");
 var type_guards_2 = require("./type-guards");
+var typescriptGenerator_1 = require("./generators/typescriptGenerator");
+var flowGenerator_1 = require("./generators/flowGenerator");
+var codecGenerator_1 = require("./generators/codecGenerator");
 var traverseArray = A.array.traverse(E.either);
+function getGenerator(_a) {
+    var type = _a.type;
+    return {
+        TypeScript: typescriptGenerator_1.typeScriptGenerator,
+        Flow: flowGenerator_1.flowGenerator,
+        CodecIoTs: codecGenerator_1.codecGenerator
+    }[type];
+}
 function getDefinitions(swagger) {
     return swagger.components
         ? E.right(swagger.components.schemas)
         : E.left(new Error("There is no definition"));
 }
 exports.getDefinitions = getDefinitions;
-var getReferenceName = function_1.flow(utils_1.replace("#/components/schemas/", ""), utils_1.toPascalCase);
-function getExactObject(options) {
-    return function_1.flow(utils_1.join(","), options.type === "TypeScript" ? utils_1.surround("{", "}") : utils_1.surround("{|", "|}"));
-}
-var escapeKey = utils_1.surround('"', '"');
-var isNumberKey = function_1.flow(parseInt, function_1.not(isNaN));
-function concatKeyAndType(key, isRequired, options) {
-    return pipeable_1.pipe(key, utils_1.doIf(function_1.constant(options.type === "Flow" && isNumberKey(key)), escapeKey), utils_1.doIf(function_1.not(function_1.constant(isRequired)), utils_1.suffix("?")), utils_1.suffix(":"), utils_1.prefix);
+function getReferenceName(options) {
+    return function_1.flow(utils_1.replace("#/components/schemas/", ""), getGenerator(options).getTypeReference);
 }
 function isRequired(key, requiredFields) {
     return requiredFields != null && requiredFields.indexOf(key) !== -1;
 }
-function getTypeUnknown(options) {
-    return options.type === "TypeScript" ? "unknown" : "mixed";
-}
 function isNullable(property) {
     return function_1.constant(property.nullable === true);
 }
-var makeTypeNullable = function_1.flow(utils_1.suffix("|null"), utils_1.surround("(", ")"));
 function fixErrorsOnProperty(property) {
     if ("allOf" in property && "type" in property) {
-        var allOf = property.allOf, otherProperties = __rest(property, ["allOf"]);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        var _a = property, allOf = _a.allOf, otherProperties = __rest(_a, ["allOf"]);
         return {
             allOf: __spreadArrays(allOf, [__assign({}, otherProperties)])
         };
     }
     if ("oneOf" in property && "type" in property) {
-        var oneOf = property.oneOf, otherProperties = __rest(property, ["oneOf"]);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        var _b = property, oneOf = _b.oneOf, otherProperties = __rest(_b, ["oneOf"]);
         return {
             oneOf: __spreadArrays(oneOf, [__assign({}, otherProperties)])
         };
     }
     if ("anyOf" in property && "type" in property) {
-        var anyOf = property.anyOf, otherProperties = __rest(property, ["anyOf"]);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        var _c = property, anyOf = _c.anyOf, otherProperties = __rest(_c, ["anyOf"]);
         return {
             anyOf: __spreadArrays(anyOf, [__assign({}, otherProperties)])
         };
@@ -94,7 +98,7 @@ function getInvalidType(options) {
     return function (property) {
         return options.exitOnInvalidType
             ? E.left(new Error("Invalid type: " + JSON.stringify(property)))
-            : E.right(getTypeUnknown(options));
+            : E.right(getGenerator(options).getTypeUnknown());
     };
 }
 function getPropertyHandler(isT, handleT) {
@@ -106,76 +110,73 @@ function getPropertyHandler(isT, handleT) {
         };
     };
 }
-var combineAllOfForTypeScript = utils_1.join("&");
-var combineAllOfForFlow = function_1.flow(utils_1.map(utils_1.prefix("...")), utils_1.join(","), utils_1.surround("{|", "|}"));
-function combineAllOf(options) {
-    return options.type === "TypeScript"
-        ? combineAllOfForTypeScript
-        : combineAllOfForFlow;
-}
 function isValidAllOf(property) {
     return (type_guards_1.isAllOf(property) &&
         property.allOf.every(function (subprop) { return type_guards_1.isReference(subprop) || type_guards_1.isObject(subprop); }));
 }
-var getTypeRef = getPropertyHandler(type_guards_1.isReference, function () { return function (property) { return E.right(getReferenceName(property.$ref)); }; });
+var getTypeRef = getPropertyHandler(type_guards_1.isReference, function (options) { return function (property) {
+    return E.right(getReferenceName(options)(property.$ref));
+}; });
 var getTypeAllOf = getPropertyHandler(isValidAllOf, function (options) { return function (property) {
-    return pipeable_1.pipe(traverseArray(property.allOf, getType(options)), E.map(combineAllOf(options)));
+    return pipeable_1.pipe(traverseArray(property.allOf, getType(options)), E.map(getGenerator(options).getTypeAllOf));
 }; });
 var getTypeOneOf = getPropertyHandler(type_guards_1.isOneOf, function (options) { return function (property) {
-    return pipeable_1.pipe(traverseArray(property.oneOf, getType(options)), E.map(utils_1.join("|")));
+    return pipeable_1.pipe(traverseArray(property.oneOf, getType(options)), E.map(getGenerator(options).getTypeOneOf));
 }; });
 var getTypeAnyOf = getPropertyHandler(type_guards_1.isAnyOf, function (options) { return function (property) {
-    return pipeable_1.pipe(traverseArray(property.anyOf, getType(options)), E.map(utils_1.join("|")));
+    return pipeable_1.pipe(traverseArray(property.anyOf, getType(options)), E.map(getGenerator(options).getTypeAnyOf));
 }; });
 var getTypeArray = getPropertyHandler(type_guards_1.isArray, function (options) { return function (property) {
-    return pipeable_1.pipe(property.items, getType(options), E.map(utils_1.surround("Array<", ">")));
+    return pipeable_1.pipe(property.items, getType(options), E.map(getGenerator(options).getTypeArray));
 }; });
-var getTypeEnum = getPropertyHandler(type_guards_1.isEnum, function () { return function (property) {
-    return E.right(pipeable_1.pipe(property.enum, utils_1.map(utils_1.surround("'", "'")), utils_1.join("|")));
+var getTypeEnum = getPropertyHandler(type_guards_1.isEnum, function (options) { return function (property) {
+    return E.right(getGenerator(options).getTypeEnum(property.enum));
 }; });
-var getTypeInteger = getPropertyHandler(type_guards_1.isInteger, function () { return function () {
-    return E.right("number");
+var getTypeInteger = getPropertyHandler(type_guards_1.isInteger, function (options) { return function () { return E.right(getGenerator(options).getTypeInteger()); }; });
+var getTypeNumber = getPropertyHandler(type_guards_1.isNumber, function (options) { return function () {
+    return E.right(getGenerator(options).getTypeNumber());
 }; });
-var getTypeNumber = getPropertyHandler(type_guards_1.isNumber, function () { return function () {
-    return E.right("number");
+var getTypeString = getPropertyHandler(type_guards_2.isString, function (options) { return function () {
+    return E.right(getGenerator(options).getTypeString());
 }; });
-var getTypeString = getPropertyHandler(type_guards_2.isString, function () { return function () {
-    return E.right("string");
-}; });
-var getTypeBoolean = getPropertyHandler(type_guards_1.isBoolean, function () { return function () {
-    return E.right("boolean");
-}; });
+var getTypeBoolean = getPropertyHandler(type_guards_1.isBoolean, function (options) { return function () { return E.right(getGenerator(options).getTypeBoolean()); }; });
 var getTypeObject = getPropertyHandler(type_guards_1.isObject, function (options) { return function (property) {
     return pipeable_1.pipe(traverseArray(Object.entries(property.properties || {}), function (_a) {
         var key = _a[0], childProperty = _a[1];
-        return pipeable_1.pipe(childProperty, getType(options), E.map(concatKeyAndType(utils_1.toCamelCase(key), isRequired(key, property.required), options)));
-    }), E.map(getExactObject(options)));
+        return pipeable_1.pipe(childProperty, getType(options), E.map(getGenerator(options).getProperty(key, isRequired(key, property.required))));
+    }), E.map(getGenerator(options).getTypeObject));
 }; });
 function getType(options) {
     return function (property) {
-        return pipeable_1.pipe(property, fixErrorsOnProperty, function_1.flow(getTypeRef(options), E.chain(getTypeAllOf(options)), E.chain(getTypeAnyOf(options)), E.chain(getTypeOneOf(options)), E.chain(getTypeArray(options)), E.chain(getTypeObject(options)), function_1.flow(E.chain(getTypeEnum(options)), E.chain(getTypeNumber(options)), E.chain(getTypeString(options)), E.chain(getTypeBoolean(options)), E.chain(getTypeInteger(options)))), E.fold(function_1.identity, getInvalidType(options)), E.map(utils_1.doIf(isNullable(property), makeTypeNullable)));
+        return pipeable_1.pipe(property, fixErrorsOnProperty, function_1.flow(getTypeRef(options), E.chain(getTypeAllOf(options)), E.chain(getTypeAnyOf(options)), E.chain(getTypeOneOf(options)), E.chain(getTypeArray(options)), E.chain(getTypeObject(options)), function_1.flow(E.chain(getTypeEnum(options)), E.chain(getTypeNumber(options)), E.chain(getTypeString(options)), E.chain(getTypeBoolean(options)), E.chain(getTypeInteger(options)))), E.fold(function_1.identity, getInvalidType(options)), E.map(utils_1.doIf(isNullable(property), getGenerator(options).makeTypeNullable)));
     };
 }
 exports.getType = getType;
-function getTypesFromSchemas(options) {
-    return function (schemas) {
-        return pipeable_1.pipe(traverseArray(Object.entries(schemas), function (_a) {
-            var key = _a[0], property = _a[1];
-            return pipeable_1.pipe(getType(options)(property), E.map(utils_1.prefix(utils_1.toPascalCase(key) + "=")));
-        }));
-    };
-}
 function checkOpenApiVersion(swagger) {
     return swagger.openapi.match(/3\.0\.\d+/)
         ? E.right(swagger)
         : E.left(new Error("Version not supported: " + swagger.openapi));
 }
-function addHeader(options) {
-    return function_1.flow(utils_1.prefix("\n"), utils_1.prefix(options.type === "TypeScript" ? '"use strict";' : "// @flow strict"));
+function getTypesFromSchemas(options) {
+    return function (schemas) {
+        return pipeable_1.pipe(traverseArray(Object.entries(schemas), function (_a) {
+            var key = _a[0], property = _a[1];
+            return pipeable_1.pipe(getType(options)(property), E.map(getGenerator(options).getTypeDefinition(key)));
+        }));
+    };
+}
+var eitherPrefix = function (b) { return function (c) {
+    return E.ap(c)(E.map(utils_1.prefix)(b));
+}; };
+function baseDefinitionsToString(options) {
+    return function_1.flow(getTypesFromSchemas(options), E.map(getGenerator(options).combineTypes));
+}
+function definitionsToString(options) {
+    return function (schemas) {
+        return pipeable_1.pipe(baseDefinitionsToString(options)(schemas), utils_1.doIf(function_1.constant(options.type === "CodecIoTs"), function_1.flow(E.map(utils_1.prefix(";")), eitherPrefix(baseDefinitionsToString(__assign(__assign({}, options), { type: "TypeScript" }))(schemas)))));
+    };
 }
 function generate(options) {
-    return function (swagger) {
-        return pipeable_1.pipe(swagger, checkOpenApiVersion, E.chain(getDefinitions), E.chain(getTypesFromSchemas(options)), E.map(function_1.flow(utils_1.map(utils_1.prefix("export type ")), utils_1.join(";"))), E.map(addHeader(options)));
-    };
+    return function_1.flow(checkOpenApiVersion, E.chain(getDefinitions), E.chain(definitionsToString(options)), E.map(getGenerator(options).addHeader));
 }
 exports.generate = generate;